requirements.txt
# BioAgent — Dependencias de producción
python-telegram-bot==22.1
google-generativeai
firebase-admin==7.2.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
APScheduler==3.11.2
sentence-transformers
numpy
PyGithub==2.8.1
python-dotenv==1.2.2
# torch se instala como dependencia de sentence-transformers
